- Variable mutation now are able to pass things as a reference, numbers are not passed by reference - Added len function, similar to pythons - Refactored garbage collector to handle cases where multiple runtime data pointers are in the GC registry, GC is now able to free them while avoiding double free errors - Small change to parser, its now able to convert large numbers to long doubles - Added a primitve get attribute feature, this is still under construction - Added Class, which are simply modeled as functions - Variables in the lookup tables in the call frames now have access modifier field, in order to enfore the access modifer for classes - Added some useful functions for working void pointers representing the actual data of a runtime object depending on its type - Added a some features to the generic hashet implementation for finding objects within a set depending on a certain predicate function - Numbers are now their object malloced seperately from the runtime object data union - Modified some of the tests

+++ parser/parser.h
@@ -96,7 +96,7 @@ typedef struct expression_component
     // all the data in this union is mutually exclusive
     union data
     {
-        double numeric_const; // NUMERIC_CONSTANT type
+        long double numeric_const; // NUMERIC_CONSTANT type
 
         char *string_literal; // NUMERIC_CONSTANT type
 
@@ -278,7 +278,7 @@ bool is_lexeme_in_list(enum token_type type, const enum token_type list[], int l
 
 bool lexeme_lists_intersect(
     enum token_type list1[], const int list1_length, enum token_type list2[], const int list2_length);
-double compute_fractional_double(Token *whole, Token *frac);
+long double compute_fractional_double(Token *whole, Token *frac);
 char *malloc_string_cpy(Parser *parser, const char *str);
 
 ExpressionComponent *malloc_expression_component(Parser *parser);
